moved pkg/normalizer
